refactor out render bundles and run compute
