+++ src/render/instantiator-webgpu.ts
@@ -1,3 +1,4 @@
+import { vec3 } from "../gl-matrix.js";
 import { assert } from "../test.js";
 import { never, capitalize, pluralize, uncapitalize } from "../util.js";
 import {
@@ -9,6 +10,7 @@ import {
   createCyDepthTexture,
   CyRndrPipeline,
   CyCompPipeline,
+  CyTexture,
 } from "./gpu-data-webgpu.js";
 import {
   PtrKind,
@@ -19,6 +21,7 @@ import {
   CyGlobal,
   CyManyBufferPtr,
   CyGlobalParam,
+  CyRndrPipelinePtr,
 } from "./gpu-registry.js";
 import { GPUBufferBindingTypeToWgslVar } from "./gpu-struct.js";
 import { createMeshPool, MeshHandle } from "./mesh-pool.js";
@@ -783,4 +786,130 @@ export function mkBindGroup(
     }),
   });
   return bindGroup;
+}
+
+export function renderBundles(
+  context: GPUCanvasContext,
+  commandEncoder: GPUCommandEncoder,
+  resources: CyResources,
+  pipelineAndBundle: [CyRndrPipeline, GPURenderBundle][],
+  backgroundColor: vec3
+) {
+  // render bundles
+  // TODO(@darzu): ordering needs to be set by outside config
+  // TODO(@darzu): same attachments need to be shared
+  let lastPipeline: CyRndrPipelinePtr | undefined;
+  let renderPassEncoder: GPURenderPassEncoder | undefined;
+  for (let [p, bundle] of pipelineAndBundle) {
+    // console.log(`rendering ${p.ptr.name}`);
+
+    if (
+      !renderPassEncoder ||
+      !lastPipeline ||
+      !isOutputEq(lastPipeline, p.ptr)
+    ) {
+      let colorAttachments: GPURenderPassColorAttachment[] = [p.ptr.output].map(
+        (o) => {
+          const isFirst = !lastPipeline;
+          const doClear = isFirst;
+          if (o.kind === "canvasTexture") return canvasAttachment(doClear);
+          else if (o.kind === "texture") {
+            let tex = resources.kindToNameToRes.texture[o.name]!;
+            return textureAttachment(tex);
+          } else {
+            never(o, "TO IMPL");
+          }
+        }
+      );
+      const depthTex =
+        resources.kindToNameToRes.depthTexture[p.ptr.depthStencil.name];
+      const depthAtt = depthTex.depthAttachment();
+
+      renderPassEncoder?.end();
+      renderPassEncoder = commandEncoder.beginRenderPass({
+        // TODO(@darzu): OUTPUT, different render targets
+        //    need different pass per different output; start with one bundle per pipeline
+        colorAttachments,
+        // TODO(@darzu): parameterize depth attachment?
+        depthStencilAttachment: depthAtt,
+      });
+    }
+
+    renderPassEncoder.executeBundles([bundle]);
+
+    lastPipeline = p.ptr;
+  }
+  renderPassEncoder?.end();
+
+  // TODO(@darzu): support multi-output
+  function isOutputEq(a: CyRndrPipelinePtr, b: CyRndrPipelinePtr) {
+    return (
+      a.output.name === b.output.name &&
+      a.depthStencil.name === b.depthStencil.name
+    );
+  }
+
+  function canvasAttachment(clear?: boolean): GPURenderPassColorAttachment {
+    return {
+      // TODO(@darzu): ANTI-ALIAS; for some reason this is tangled in AA?
+      // view: canvasTextureView!,
+      // resolveTarget: context.getCurrentTexture().createView(),
+      view: context.getCurrentTexture().createView(),
+      // TODO(@darzu): is this how we want to handle load vs clear?
+      loadOp: clear ? "clear" : "load",
+      clearValue: {
+        r: backgroundColor[0],
+        g: backgroundColor[1],
+        b: backgroundColor[2],
+        a: 1,
+      },
+      storeOp: "store",
+    };
+  }
+
+  // TODO(@darzu): move into CyTexture
+  function textureAttachment(tex: CyTexture): GPURenderPassColorAttachment {
+    // TODO(@darzu): parameterizable?
+    return {
+      view: tex.texture.createView(),
+      // loadOp: "load",
+      // TODO(@darzu): handle load vs clear
+      loadOp: "clear",
+      clearValue: {
+        r: backgroundColor[0],
+        g: backgroundColor[1],
+        b: backgroundColor[2],
+        a: 1,
+      },
+      storeOp: "store",
+    };
+  }
+}
+
+export function doCompute(
+  device: GPUDevice,
+  resources: CyResources,
+  commandEncoder: GPUCommandEncoder,
+  pipeline: CyCompPipeline
+) {
+  const compPassEncoder = commandEncoder.beginComputePass();
+  compPassEncoder.setPipeline(pipeline.pipeline);
+
+  // TODO(@darzu): de-dupe?
+  const resBindGroupLayout = pipeline.bindGroupLayout;
+  const resUsages = normalizeResources(pipeline.ptr.resources);
+  const resBindGroup = mkBindGroup(
+    device,
+    resources,
+    resBindGroupLayout,
+    resUsages,
+    "many"
+  );
+
+  compPassEncoder.setBindGroup(0, resBindGroup);
+  // TODO(@darzu): parameterize workgroup count
+  compPassEncoder.dispatchWorkgroups(
+    ...(pipeline.ptr.workgroupCounts ?? [1, 1, 1])
+  );
+  compPassEncoder.end();
 }
